+++ src/main/java/cn/huanzi/qch/baseadmin/ccb/channel/controller/ChannelController.java
@@ -1,10 +1,8 @@
 package cn.huanzi.qch.baseadmin.ccb.channel.controller;
 
-import cn.huanzi.qch.baseadmin.ccb.channel.repository.ChannelRepository;
-import cn.huanzi.qch.baseadmin.common.controller.CommonController;
+import cn.huanzi.qch.baseadmin.ccb.upload.service.FileService;
 import cn.huanzi.qch.baseadmin.ccb.channel.pojo.Channel;
 import cn.huanzi.qch.baseadmin.ccb.channel.service.ChannelService;
-import cn.huanzi.qch.baseadmin.ccb.channel.vo.ChannelVo;
 import cn.huanzi.qch.baseadmin.common.pojo.Result;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -14,8 +12,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
 
-import javax.swing.event.ChangeEvent;
-
 /**
  *  Controller
  *
@@ -24,13 +20,30 @@ import javax.swing.event.ChangeEvent;
  */
 @RestController
 @RequestMapping("/ccb/channel/")
-public class ChannelController extends CommonController<ChannelVo, Channel, Integer> {
+public class ChannelController {
     @Autowired
     private ChannelService channelService;
 
+    @Autowired
+    private FileService fileService;
+
     @GetMapping("index")
     public ModelAndView index() {
-        return new ModelAndView("ccb/channel/index");
+        return new ModelAndView("ccb/channel/channel");
+    }
+
+    @GetMapping("channelForm")
+    public ModelAndView channelForm(Integer id) {
+        ModelAndView mav = new ModelAndView("ccb/channel/channelForm");
+
+        if (id == null) {
+            mav.addObject("channel", null);
+        } else {
+            Channel channel = channelService.getById(id);
+            mav.addObject("channel", channel);
+        }
+
+        return mav;
     }
 
     @GetMapping("page")
@@ -69,8 +82,21 @@ public class ChannelController extends CommonController<ChannelVo, Channel, Inte
 
     @PostMapping("delete")
     public Result deleteById(Integer id) {
-        channelService.deleteById(id);
-        return Result.of(id, true, 200, "success");
+        if (id == null) {
+            return Result.of(id, false, 400, "未输入id");
+        }
+        Channel result = channelService.getById(id);
+        if (result == null) {
+            Result.of(id, false, 400, "没有该记录");
+        }
+        try {
+            fileService.deleteImage(result.getIcon());
+            fileService.deleteImage(result.getImage());
+            channelService.deleteById(id);
+        } catch (Exception e) {
+            return Result.of(id, false, 400, "删除失败");
+        }
+        return Result.of(id, true, "删除成功");
     }
 }
